Gracefully handle connection timeouts for registrars that point to non-functional WHOIS servers
This fixes an issue see with the `pairdomains.com` registrar, whose WHOIS record terminate with a record that includes this:

```
    Registrar WHOIS Server: pairdomains.com
```

The actual WHOIS server for the registrar is `whois.pairdomains.com`, so prior to this patch, `Whois.lookup/1` would fail after 60 seconds with `{:error, :timed_out}`. Now, we'll fall back to the most recent successfully retrieved record after `:connect_timeout` (10 seconds by default).